refactor(vault): split DepositProgressView and address review feedback
- Split monolithic DepositProgressView.tsx into per-component files under
  DepositProgressView/ (one component per file per CLAUDE.md).
- Handle DepositFlowStep.COMPLETED in getVisualStep so the pill shows
  11 of 11 instead of falling through to the default 1.
- Extract TOTAL_VISUAL_STEPS to steps.ts so the constant lives next to
  the step mapping it gates.
- Add BtcConfirmationDetail panel surfacing started-at, est. remaining
  countdown, and Bitcoin TX link inside the AWAIT_BTC_CONFIRMATION
  active row.
- Track AWAIT_BTC_CONFIRMATION entry time and pegin tx hash in
  useDepositFlow so the panel has the data it needs.
