+++ services/vault/src/components/simple/DepositProgressView/ProgressBar.tsx
@@ -1,0 +1,21 @@
+interface ProgressBarProps {
+  percent: number;
+}
+
+export function ProgressBar({ percent }: ProgressBarProps) {
+  const clamped = Math.max(0, Math.min(1, percent));
+  return (
+    <div
+      role="progressbar"
+      aria-valuemin={0}
+      aria-valuemax={100}
+      aria-valuenow={Math.round(clamped * 100)}
+      className="h-1 w-full overflow-hidden rounded-full bg-secondary-strokeLight"
+    >
+      <div
+        className="h-full bg-success-light transition-[width] duration-300"
+        style={{ width: `${clamped * 100}%` }}
+      />
+    </div>
+  );
+}
