+++ services/vault/src/components/simple/DepositProgressView/StepConnector.tsx
@@ -1,0 +1,5 @@
+export function StepConnector() {
+  return (
+    <div className="ml-[15px] h-6 w-0 border-l-2 border-solid border-secondary-strokeDark" />
+  );
+}
