+++ services/vault/src/components/simple/DepositProgressView/index.ts
@@ -1,0 +1,6 @@
+export {
+  DepositProgressView,
+  type BtcConfirmationDetailData,
+  type DepositProgressViewProps,
+} from "./DepositProgressView";
+export { buildStepItems, getVisualStep } from "./steps";
